Learnings updated:
Learnt how to adapt a react app to make use of API infrastructure to pull images from another web application using a query hook.

Learnt correct way to structure files across react directorires. It is good practice to keep the style and basic css files within the same directory as the index.html file. This simplifies the refrencing and linking of css files within index.html

I learnt the difference bewteen a website and web app.

The version of the react build should correspond to the versions specified in the package.json config file for react and reatc dom. Even though react generally is backward compatible - however, some web applications are finicky when it comes to dependencies.

Learnt how to display name of web app in home page.

Uploaded image to a CDN (Sirv) to host images and then use the link as a css variable to reference that variable as a way to change images of backgrounds and even borders.

The same background image was use for the border image. I choose this to have a consistent them with the web app. Using images for borders is similar to applying materials to game assets or graphical objects.